requirements.txt

# bitstring
# pytest
LatLon23

# fdxdecode.py, dumpserial.py
pyserial==3.0.1
